requirements.txt
ultralytics
opencv-python-headless
streamlit
pillow
numpy
